==> 07-conclusion.Rmd <==
# Conclusion

After conducting the study on stock data, Reddit post data and Google search related data, we have a better understanding of the WallStreetBets. Besides answering the 3 questions that we asked in [Introduction], we also have some other interesting findings. 

Here are the answers of the questions that we asked in [Introduction]:

1. How big is the impact of r/wallstreetbets subreddit group?

The impact of the r/wallstreetbets subreddit group is most prominent during January and February, 2021 and there is relatively smaller but still noticeable impact during the 2nd wave during May to July, 2021. However, besides these 2 periods, there is not much impact of r/wallstreetbets subreddit group on the stock price and trading volume as well as the general public.

2. For each targeted stock, what’s the relationship between Reddit posts and stock price/volume etc?

We find the top 10 stocks that are most mentioned in the Reddit posts, which are also the target stocks. Generally there is no obvious relationship between the post count and daily price change for both `AMC` and `GME`. Because the majority of the posts are about `AMC` and `GME`, we can infer the influence is even less for the rest of the posts discussing about other stocks, therefore, we can conclude that there is no relationship between the posts count and daily price change for all those stocks.

On the other side, as for the relationship between Reddit posts and volume, we find some weak correlation between posts counts and volume, the bigger the post count is, the higher the volume is. With the above two observations, we can infer that the sentiments in the posts can be positive or negative, so people may trade in either directions. But the discussions happened in r/wallstreetbets definitely attracted people's attentions to those stocks, causing much higher trading volume in the time period.

3. Why does r/wallstreetbets choose these target stocks? What’re their similarities? How do their choices change as time evolves?

The redditers in r/wallstreetbets choose these target stocks because these stocks are heavily shorted by big financial institutes. There is a change in the target stocks between the 2 waves. While `PLTR` is the stock that is only active in 1st wave, there are other stocks that are active only in the 2nd wave, which are `CLOV`, `SND`, `RETA` respectively. For `GME`, `AMC`, `BB`, `NOK`, `NAKD`, they are active in both of the waves. For the stocks that are active in the 1st wave, their market capital were driven to a high level at the end of the wave. As a result, it is hard to drive the stock price higher in the 2nd wave with limited capital and time, therefore, they started to look at other target stocks.

Besides all the answers to the above question, here are some of interesting findings from all the graph we plot.

- `GME` is the focus of the 1st WSB wave while `AMC` is the 2nd WSB wave.

- The risk of those WSB target stocks is extremely high during the 2 wave period, with the moving standard deviation be to be 100 times as comparing to the normal level.

- `NAKD` is an outlier in the 1st WSB wave even if it is one of the top 10 most mention word.

- There are huge amount of new users joining r/wallstreetbets daily at the end of the January. However, after that, there is not much new joiners afterwards, even in the 2nd wave. We can infer that it is the same group of people who contribute to the 1st WSB wave and 2nd WSB wave.

- The overall sentiments of posts in the 1st WSB wave is slightly more negative than that is in the 2nd WSB wave.

- The sentiment in the r/wallstreetbets subreddit group can be quickly spread across different people and that will amplify the change in stock price since there are a large amount of people buying and selling the same group of stocks at the same time.

- The 1st wave of WSB caught more public attention than the 2nd wave of WSB.

Despite of all these findings, our analysis has some limitations. 

Firstly, for the Reddit post data, since quite a lot of posts are removed or deleted by the time we downloaded from the data source, therefore, our study using the Reddit posts may not reflect the true situation during the 1st and 2nd WSB wave, which means some of the conclusions may be slightly different from what actually happening at that point of time.

Secondly, for the Google search data, since we cannot get the absolute search count for the Google search but rather, we can only get the relative score provided by Google. Also, for the keyword selection, we could also include more words in the data source to find more interesting facts.

Therefore, for future improvements, we could try to find the data source that supplies more comprehensive Reddit posts and include more keywords in Google search data. That will help to disclose more information.
